backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
Deprecated==1.2.14
Django==4.2.3
django-cors-headers==4.2.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
idna==3.4
jwcrypto==1.5.0
oauthlib==3.2.2
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3
wrapt==1.15.0
